+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# Fixed
+- Regression around handling of "signature-less" builtin functions.
+  A ValueError will no-longer be raised in these cases.
+
 
 ## [0.1.0] - 2020-02-25